machdep/darwin: add missing doload.c, classproc.c, and generated Makefiles; fixes clean-clone build failure
FossilOrigin-Name: 486d71704f1b8b1bcea1ff993ddbc0bedb23327b96ed5cc4f3c0c14fd9885134

+++ src/atk/eq/eq.c
@@ -209,7 +209,7 @@ long pos;
     }
 
     self->gap += 1;
-    mark_UpdateMarks(self->markList, pos , -1);
+    mark_UpdateMarks(self->markList, pos , (long)-1);
 }
 
 /*
